+++ include/BDSOutputROOT.hh
@@ -50,6 +50,9 @@ public:
   virtual void NewFile();    ///< Open a new file.
   virtual void CloseFile();  ///< Write contents and close file.
 
+#ifdef SIXTRACKLINK
+  virtual void UpdateSamplers();
+#endif
 private:
   /// Copy header and write to file.
   virtual void WriteHeader();