+++ include/BDSTransform3D.hh
@@ -47,11 +47,11 @@ public:
   
   BDSTransform3D(const G4String& aName,
                  G4double x,
-		             G4double y,
-		             G4double z,
-		             G4double phi,
-		             G4double theta,
-		             G4double psi);
+		 G4double y,
+		 G4double z,
+		 G4double phi,
+		 G4double theta,
+		 G4double psi);
   
   BDSTransform3D(const G4String& nameIn,
                  G4double x,
@@ -68,9 +68,9 @@ public:
   G4double dx;
   G4double dy;
   G4double dz;
+  G4double dPhi;
   G4double dTheta;
   G4double dPsi;
-  G4double dPhi;
   G4double axisX;
   G4double axisY;
   G4double axisZ;