+++ src/BDSComponentFactory.cc
@@ -1162,7 +1162,8 @@ BDSAcceleratorComponent* BDSComponentFactory::CreateElement()
 			 element->geometryFile,
 			 element->angle * CLHEP::rad,
 			 &vacuumBiasVolumeNames,
-			 element->autoColour));
+			 element->autoColour,
+			 element->markAsCollimator));
 }
 
 BDSAcceleratorComponent* BDSComponentFactory::CreateSolenoid()
@@ -1732,9 +1733,9 @@ BDSAcceleratorComponent* BDSComponentFactory::CreateTransform3D()
                               element->xdir * CLHEP::m,
                               element->ydir * CLHEP::m,
                               element->zdir * CLHEP::m,
-                              element->phi * CLHEP::rad,
+                              element->phi   * CLHEP::rad,
                               element->theta * CLHEP::rad,
-                              element->psi * CLHEP::rad);
+                              element->psi   * CLHEP::rad);
   }
 }