+++ src/BDSOutput.cc
@@ -30,6 +30,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "BDSHitEnergyDeposition.hh"
 #include "BDSHitEnergyDepositionGlobal.hh"
 #include "BDSHitSampler.hh"
+#include "BDSHitSamplerLink.hh"
 #include "BDSOutput.hh"
 #include "BDSOutputROOTEventAperture.hh"
 #include "BDSOutputROOTEventBeam.hh"
@@ -279,6 +280,7 @@ void BDSOutput::FillEvent(const BDSEventInfo*                            info,
 			  const G4PrimaryVertex*                         vertex,
 			  const BDSHitsCollectionSampler*                samplerHitsPlane,
 			  const BDSHitsCollectionSampler*                samplerHitsCylinder,
+                          const BDSHitsCollectionSamplerLink*            samplerHitsLink,
 			  const BDSHitsCollectionEnergyDeposition*       energyLoss,
 			  const BDSHitsCollectionEnergyDeposition*       energyLossFull,
 			  const BDSHitsCollectionEnergyDeposition*       energyLossVacuum,
@@ -313,6 +315,8 @@ void BDSOutput::FillEvent(const BDSEventInfo*                            info,
     {FillSamplerHits(samplerHitsPlane, BDSOutput::HitsType::plane);}
   if (samplerHitsCylinder)
     {FillSamplerHits(samplerHitsCylinder, BDSOutput::HitsType::cylinder);}
+  if (samplerHitsLink)
+    {FillSamplerHitsLink(samplerHitsLink);}
   if (energyLoss)
     {FillEnergyLoss(energyLoss,        BDSOutput::LossType::energy);}
   if (energyLossFull)
@@ -671,6 +675,29 @@ void BDSOutput::FillSamplerHits(const BDSHitsCollectionSampler* hits,
     }
 }
 
+void BDSOutput::FillSamplerHitsLink(const BDSHitsCollectionSamplerLink* hits)
+{
+  G4int nHits = hits->entries();
+  if (nHits == 0) // integer so ok to compare
+    {return;}
+  for (int i = 0; i < (int)hits->entries(); i++)
+    {
+      const BDSHitSamplerLink* hit = (*hits)[i];
+      G4int samplerID = hit->samplerID;
+      samplerID += 1; // offset index by one due to primary branch.
+      samplerTrees[samplerID]->Fill(hit, storeSamplerMass, storeSamplerCharge, storeSamplerPolarCoords, storeSamplerIon, storeSamplerRigidity, storeSamplerKineticEnergy);
+    }
+  // extra information - do only once at the end
+  G4bool firstSampler = true;
+  for (auto& sampler : samplerTrees)
+    {
+      if (firstSampler) // skip primaries (1st sampler) as it always has extras filled in
+	{firstSampler = false; continue;}
+      if (storeSamplerIon)
+	{sampler->FillIon();}
+    }
+}
+
 void BDSOutput::FillEnergyLoss(const BDSHitsCollectionEnergyDepositionGlobal* hits,
 			       const LossType lossType)
 {
@@ -684,7 +711,7 @@ void BDSOutput::FillEnergyLoss(const BDSHitsCollectionEnergyDepositionGlobal* hi
       {return; break;} // don't fill for other types of hits
     }
 
-  G4int nHits = hits->entries();
+  G4int nHits = (G4int)hits->entries();
   if (nHits == 0) // integer so ok to compare
     {return;}
   switch (lossType)
@@ -731,7 +758,7 @@ void BDSOutput::FillEnergyLoss(const BDSHitsCollectionEnergyDepositionGlobal* hi
 void BDSOutput::FillEnergyLoss(const BDSHitsCollectionEnergyDeposition* hits,
 			       const LossType lossType)
 {
-  G4int nHits            = hits->entries();
+  G4int nHits            = (G4int)hits->entries();
   if (nHits == 0)
     {return;}
   G4int indELoss         = histIndices1D["Eloss"];
@@ -871,7 +898,7 @@ void BDSOutput::FillTrajectories(const BDSTrajectoriesToStore* trajectories)
 void BDSOutput::FillCollimatorHits(const BDSHitsCollectionCollimator* hits,
 				   const std::vector<const BDSTrajectoryPointHit*>& primaryLossPoints)
 {
-  G4int nHits = hits->entries();
+  G4int nHits = (G4int)hits->entries();
   for (G4int i = 0; i < nHits; i++)
     {
       BDSHitCollimator* hit = (*hits)[i];
@@ -933,7 +960,7 @@ void BDSOutput::FillApertureImpacts(const BDSHitsCollectionApertureImpacts* hits
     {return;}
 
   G4int nPrimaryImpacts = 0;
-  G4int nHits = hits->entries();
+  G4int nHits = (G4int)hits->entries();
   G4int histIndex = histIndices1D["PFirstAI"];
   for (G4int i = 0; i < nHits; i++)
     {