+++ src/BDSTrajectoryPoint.cc
@@ -35,7 +35,6 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "globals.hh"
 #include "G4Allocator.hh"
-#include "G4NavigationHistory.hh"
 #include "G4ProcessType.hh"
 #include "G4Step.hh"
 #include "G4ThreeVector.hh"
@@ -46,6 +45,8 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include <ostream>
 
 G4Allocator<BDSTrajectoryPoint> bdsTrajectoryPointAllocator;
+
+G4double BDSTrajectoryPoint::dEThresholdForScattering = 1e-8;
 
 // Don't use transform caching in the aux navigator as it's used for all over the geometry here.
 BDSAuxiliaryNavigator* BDSTrajectoryPoint::auxNavigator = new BDSAuxiliaryNavigator();
@@ -247,7 +248,7 @@ void BDSTrajectoryPoint::StoreExtrasLink(const G4Track* track)
   G4double rigidity = 0;
   if (BDS::IsFinite(charge))
     {rigidity = BDS::Rigidity(track->GetMomentum().mag(), charge);}
-  extraLink = new BDSTrajectoryPointLink(charge,
+  extraLink = new BDSTrajectoryPointLink((G4int)charge,
 					 BDSGlobalConstants::Instance()->TurnsTaken(),
 					 dynamicParticleDef->GetMass(),
 					 rigidity);
@@ -312,6 +313,12 @@ G4bool BDSTrajectoryPoint::IsScatteringPoint(const G4Step* step)
 {
   const G4StepPoint* postPoint   = step->GetPostStepPoint();
   const G4VProcess*  postProcess = postPoint->GetProcessDefinedStep();
+  
+  G4double totalEnergyDeposit = step->GetTotalEnergyDeposit();
+  
+  G4Track* t = step->GetTrack();
+  if (t->GetCurrentStepNumber() == 1 && t->GetStepLength() < 1e-5 && totalEnergyDeposit < 1e-5)
+    {return false;} // ignore the first really small step
 
   G4int postProcessType    = -1;
   G4int postProcessSubType = -1;
@@ -320,8 +327,7 @@ G4bool BDSTrajectoryPoint::IsScatteringPoint(const G4Step* step)
       postProcessType    = postProcess->GetProcessType();
       postProcessSubType = postProcess->GetProcessSubType();
     }
-
-  G4double totalEnergyDeposit = step->GetTotalEnergyDeposit();
+  
   return BDSTrajectoryPoint::IsScatteringPoint(postProcessType, postProcessSubType, totalEnergyDeposit);
 }
   
@@ -338,7 +344,7 @@ G4bool BDSTrajectoryPoint::IsScatteringPoint(G4int postProcessType,
   G4bool notUndefined      = postProcessType != G4ProcessType::fNotDefined; // for crystal channelling
 
   // energy can change in transportation step (EM)
-  if (totalEnergyDeposit > 1e-9)
+  if (totalEnergyDeposit > dEThresholdForScattering)
     {return true;}
 
   return initialised && notTransportation && notGeneral && notParallel && notUndefined;